Expanding tooltipSettings when market segment was chosen
