+++ src/ducks/SANCharts/data.js
@@ -2,6 +2,7 @@ import React from 'react'
 import { Line, Bar, Area } from 'recharts'
 import { usdFormatter } from './utils'
 import { millify } from '../../utils/formatting'
+import { getDateFormats, getTimeFormats } from '../../utils/dates'
 
 export const Events = {
   trendPositionHistory: {
@@ -365,3 +366,45 @@ export const compatabilityMap = {
 export const SOCIAL_SIDEBAR = 'SOCIAL_SIDEBAR'
 export const ASSETS_SIDEBAR = 'ASSETS_SIDEBAR'
 export const HISTOGRAM_SIDEBAR = 'HISTOGRAM_SIDEBAR'
+
+const LARGE_NUMBER_THRESHOLD = 99999
+
+const FORMATTER = value => {
+  if (!value && typeof value !== 'number') {
+    return 'No data'
+  }
+
+  if (value > LARGE_NUMBER_THRESHOLD) {
+    return millify(value, 2)
+  }
+
+  return Number.isInteger(value) ? value : value.toFixed(2)
+}
+
+export const tooltipSettings = {
+  datetime: {
+    formatter: value => {
+      const date = new Date(value)
+      const { HH, mm } = getTimeFormats(date)
+      const { MMMM, DD, YYYY } = getDateFormats(date)
+      return `${HH}:${mm}, ${MMMM} ${DD}, ${YYYY}`
+    }
+  },
+  isAnomaly: {
+    label: 'Anomaly',
+    formatter: v => v
+  },
+  trendingPosition: {
+    label: 'Trending Position',
+    formatter: ([val]) => Events.position.formatter(val)
+  }
+}
+
+Object.values(Metrics).forEach(
+  ({ key, dataKey = key, formatter = FORMATTER, label }) => {
+    tooltipSettings[dataKey] = {
+      label,
+      formatter
+    }
+  }
+)
